pyproject.toml
[tool.poetry]
name = "gptforum"
version = "0.1.0"
description = ""
authors = ["KirillPoke"]
license = "AGPL"
readme = "README.md"
package-mode = false
[tool.poetry.dependencies]
python = ">3.10, <3.12"
Django = "5.0.1"
djangorestframework = "3.14.0"
openai = "0.27.8"
django-cors-headers = "4.2.0"
psycopg2 = "2.9.7"
random-username = "1.0.2"
django-tree-queries = "^0.15.0"
daphne = "^4.0.0"
twisted = { extras = ["http2", "tls"], version = "^23.8.0" }
msgpack = "^1.0.7"
whitenoise = "^6.6.0"
django-q2 = "^1.6.1"
dj-url-filter = "^0.4.4"
boto3 = "^1.34.41"
authlib = "^1.3.0"
requests = "^2.31.0"
drf-jwt = "^1.19.2"
django-request-logging = "^0.7.5"
pyinstrument = "^5.0.0"
django-silk = "^5.3.2"
django-auto-prefetching = "^0.2.12"

[tool.poetry.group.dev.dependencies]
pre-commit = "3.3.3"
poetry = "^1.8.3"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
